Reinstate sending mails per smtp.
We have had weird issues with the call to mail failing, and the server
is running exim4 anyways, so might as well.

I am a bit uneasy about this, since I am not sure why it was changes to
use `mail`, but this is the best I can do for now.